pyproject.toml
[tool.poetry]
name = "bifrost-rpc"
version = "0.6.4"
description = ""
license = "MIT"
authors = []
packages = [
    { include = "bifrostrpc" }
]
# make sure py.typed files are included when installing via poetry (setup.py does this differently)
include = ["py.typed"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
paradox-codegen = { git = "https://github.com/phodge/paradox.git", tag = "0.18.3" }

[tool.poetry.dev-dependencies]
mypy = "^0.901"
flake8 = "^3.8.4"
pytest = "^6.0.0"
pylint = "^2.6.0"
# these are required for pylint to run
pytoml = "^0.1.21"
Flask = "^2.2.2"
types-dataclasses = "^0.6.1"
types-Flask = "^1.1.5"
requests = "^2.28.1"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
